api-gateway/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
huggingface_hub==0.16.4
sentence-transformers==2.2.2
qdrant-client==1.6.0
gunicorn==21.2.0 